requirements.txt
# Django Inventory System - Simplified
# Core
Django==4.2.11
python-dotenv==1.2.1

# Database
sqlparse==0.4.4

# Import/Export
django-import-export>=3.0.0
openpyxl>=3.0.0

# Utilities
pillow>=10.0.0
